Update wastelandponygen.html
Update makes the clothing generation into a dynamic system capable of creating many unique outfits. Also adds a faction selection system, which can be used to generate appropriate equipment for the type of character in question.

Also added an inventory section which tracks the number of and weight of everything the character has.

Currently the armor and weapon generation systems are offline. They still exist, but weapon creation was tangled up in the old outfit generation system.